Move preview classes to separate file for autoloading reloading
