+++ spec/components/find/deadline_banner_component_preview.rb
@@ -29,38 +29,4 @@ module Find
       )
     end
   end
-
-  class ShowApplyDeadlineBannerCycleTimetable < CycleTimetable
-    def self.show_apply_deadline_banner?
-      true
-    end
-
-    def self.show_cycle_closed_banner?
-      false
-    end
-  end
-
-  class ShowCycleClosedBannerTimetable < CycleTimetable
-    def self.show_apply_deadline_banner?
-      false
-    end
-
-    def self.show_cycle_closed_banner?
-      true
-    end
-  end
-
-  class ShowApplyOpensSoonBannerTimetable < CycleTimetable
-    def self.show_apply_deadline_banner?
-      false
-    end
-
-    def self.show_cycle_closed_banner?
-      false
-    end
-
-    def self.show_apply_opens_soon_banner?
-      true
-    end
-  end
 end
